fix(ultrasonic): use NewPing lib

+++ arduino/lib/helpers.h
@@ -6,12 +6,12 @@
 class Helpers
 {
 public:
-  int distanceToLevel(float distance)
+  int distanceToLevel(unsigned long distance)
   {
     return constrain(map(distance, DISTANCE_FOR_EMPTY, DISTANCE_FOR_FULL, 0, 100), 0, 100);
   }
 
-  String arrayToString(float array[], byte len)
+  String arrayToString(unsigned long array[], byte len)
   {
     String str;
     for (byte i = 0; i < len; i++)
